Integrate CommandSyntaxWords class into CliSyntax.java
